# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A social network prove XD",
    "version": "0.2.0",
    "title": "social_prove",
    "contact": {
      "email": "erickzuria71@gmail.com"
    },
    "license": {
      "name": "CC",
      "url": ""
    }
  },
  "host": "",
  "basePath": "",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/posts": {
      "get": {
        "tags": [
          "user",
          "admin",
          "super"
        ],
        "description": "Returns a list of posts.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit to pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetPostsResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "user"
        ],
        "description": "Register a new post.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RegisterPostRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "user",
          "admin",
          "super"
        ],
        "description": "Returns a post by id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetPostResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "user",
          "admin",
          "super"
        ],
        "description": "Update a post by id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UpdatePostRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user",
          "admin",
          "super"
        ],
        "description": "Delete a post by id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/posts/{id}/like": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Add like to a post by id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "description": "Remove like from a post by id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/comments/post/{id}": {
      "get": {
        "tags": [
          "user",
          "admin",
          "super"
        ],
        "description": "Returns a list of comments by post id.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Access token.",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page to pagination",
            "required": false,
            "type": "number"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit to pagination",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCommentsResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorBadRequest"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorUnauthorized"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorInternalServerError"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "AccessToken": {
      "type": "apiKey",
      "name": "access_token",
      "in": "header"
    }
  },
  "definitions": {
...
    "ErrorBadRequest": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Bad resquest"
        }
      }
    },
    "ErrorUnauthorized": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Unauthorized"
        }
      }
    },
    "ErrorInternalServerError": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Interal server error"
        }
      }
    },
...
    "PostUser": {
      "type": "object",
      "required": [
        "first_name",
        "id",
        "last_name",
        "picture"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "5d7273cb40d82abd58b2a8ef"
        },
        "first_name": {
          "type": "string",
          "example": "John",
          "default": ""
        },
        "last_name": {
          "type": "string",
          "example": "Doe",
          "default": ""
        },
        "picture": {
          "type": "string",
          "format": "url",
          "example": "http://example.com/me.jpg"
        }
      }
    },
...
    "RegisterPostRequest": {
      "type": "object",
      "required": [
        "badge",
        "pictures",
        "user_id"
      ],
      "properties": {
        "user_id": {
          "type": "string",
          "example": "5d7273cb40d82abd58b2a8ef"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "array",
          "example": [
            8.2,
            -12.5
          ],
          "items": {
            "type": "number",
            "example": -12.5
          }
        },
        "pictures": {
          "type": "array",
          "example": [
            "http://example.com/cover.jpg",
            "http://example.com/image.jpg"
          ],
          "items": {
            "type": "string",
            "example": "http://example.com/cover.jpg"
          }
        },
        "badge": {
          "type": "string",
          "example": "10.4 KM"
        }
      }
    },
    "PostResponse": {
      "type": "object",
      "required": [
        "badge",
        "id",
        "pictures",
        "user"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "5d7273cb40d82abd58b2a8ef"
        },
        "user": {
          "$ref": "#/definitions/PostUser"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "array",
          "example": [
            8.2,
            -12.5
          ],
          "items": {
            "type": "number",
            "example": -12.5
          }
        },
        "pictures": {
          "type": "array",
          "example": [
            "http://example.com/cover.jpg",
            "http://example.com/image.jpg"
          ],
          "items": {
            "type": "string",
            "example": "http://example.com/cover.jpg"
          }
        },
        "badge": {
          "type": "string",
          "example": "10.4 KM"
        },
        "likes": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "5d7273cb40d82abd58b2a8ef"
          }
        },
        "created_at": {
          "type": "string",
          "format": "datetime",
          "example": "2019-12-11T19:22:49.503Z"
        },
        "updated_at": {
          "type": "string",
          "format": "datetime",
          "example": "2019-12-11T19:22:49.503Z"
        }
      }
    },
    "GetPostResponse": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/PostResponse"
        }
      }
    },
    "GetPostsResponse": {
      "type": "object",
      "properties": {
        "posts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PostResponse"
          }
        },
        "total": {
          "type": "number",
          "format": "uint8"
        }
      }
    },
    "UpdatePostRequest": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "pictures": {
          "type": "array",
          "example": [
            "http://example.com/cover.jpg",
            "http://example.com/image.jpg"
          ],
          "items": {
            "type": "string",
            "example": "http://example.com/cover.jpg"
          }
        },
        "badge": {
          "type": "string",
          "example": "10.4 KM"
        }
      }
    },
...
    "CommentResponse": {
      "type": "object",
      "required": [
        "body",
        "id"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "5d7273cb40d82abd58b2a8ef"
        },
        "post_id": {
          "type": "string",
          "example": "5d7273cb40d82abd58b2a8ef"
        },
        "user": {
          "$ref": "#/definitions/PostUser"
        },
        "body": {
          "type": "string"
        },
        "location": {
          "type": "array",
          "example": [
            8.2,
            -12.5
          ],
          "items": {
            "type": "number",
            "example": -12.5
          }
        },
        "likes": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "5d7273cb40d82abd58b2a8ef"
          }
        },
        "created_at": {
          "type": "string",
          "format": "datetime",
          "example": "2019-12-11T19:22:49.503Z"
        },
        "updated_at": {
          "type": "string",
          "format": "datetime",
          "example": "2019-12-11T19:22:49.503Z"
        }
      }
    },
...
    "GetCommentsResponse": {
      "type": "object",
      "properties": {
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CommentResponse"
          }
        },
        "total": {
          "type": "number",
          "format": "uint8"
        }
      }
    },
...
  }
}